removed responseBloc, added a few bug fixes, renamed and moved some folders
