+++ lib/main.dart
@@ -17,9 +17,6 @@ import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:graphql_flutter/graphql_flutter.dart';
-import 'package:provider/provider.dart';
-
-import 'bloc/redeemBloc/redeem_bloc.dart';
 
 final _dio = Dio();
 final _mlbApi = MLBAthleteAPI(_dio);
@@ -32,37 +29,32 @@ void main() async {
   final _getPairInfoUseCase = GetPairInfoUseCase(_subGraphRepo);
   final _getSwapInfoUseCase = GetSwapInfoUseCase(_getPairInfoUseCase);
   print("Graph QL CLient initialized}");
-  runApp(MultiProvider(
-    providers: [
-      Provider<RedeemBloc>(create: (_) => RedeemBloc()),
-    ],
-    child: GraphQLProvider(
-      client: _gQLClient,
-      child: MultiRepositoryProvider(
-        providers: [
-          RepositoryProvider(create: (context) => _subGraphRepo),
-          RepositoryProvider(
-            create: (context) => MLBRepo(_mlbApi),
-          ),
-          RepositoryProvider(
-            create: (context) => NFLRepo(),
-          ),
-          RepositoryProvider(
-            create: (context) => CoinGeckoRepo(_coinGeckoApi),
-          ),
-          RepositoryProvider(create: (context) => _getPairInfoUseCase),
-          RepositoryProvider(create: (context) => _getSwapInfoUseCase),
-          RepositoryProvider(
-              create: (context) => GetBuyInfoUseCase(_getSwapInfoUseCase)),
-          RepositoryProvider(
-              create: (context) => GetSellInfoUseCase(_getSwapInfoUseCase)),
-          RepositoryProvider(
-              create: (context) => GetPoolInfoUseCase(_getPairInfoUseCase)),
-          RepositoryProvider(
-              create: (context) => GetAllLiquidityInfoUseCase(_subGraphRepo)),
-        ],
-        child: MyApp(),
-      ),
+  runApp(GraphQLProvider(
+    client: _gQLClient,
+    child: MultiRepositoryProvider(
+      providers: [
+        RepositoryProvider(create: (context) => _subGraphRepo),
+        RepositoryProvider(
+          create: (context) => MLBRepo(_mlbApi),
+        ),
+        RepositoryProvider(
+          create: (context) => NFLRepo(),
+        ),
+        RepositoryProvider(
+          create: (context) => CoinGeckoRepo(_coinGeckoApi),
+        ),
+        RepositoryProvider(create: (context) => _getPairInfoUseCase),
+        RepositoryProvider(create: (context) => _getSwapInfoUseCase),
+        RepositoryProvider(
+            create: (context) => GetBuyInfoUseCase(_getSwapInfoUseCase)),
+        RepositoryProvider(
+            create: (context) => GetSellInfoUseCase(_getSwapInfoUseCase)),
+        RepositoryProvider(
+            create: (context) => GetPoolInfoUseCase(_getPairInfoUseCase)),
+        RepositoryProvider(
+            create: (context) => GetAllLiquidityInfoUseCase(_subGraphRepo)),
+      ],
+      child: MyApp(),
     ),
   ));
 }
